refactor: replace u8/u16/u32/u64 types with uint8/uint16/int32/int64 for consistency; update related code and namespaces

+++ modules/core/core/public/core/libc/mem.h
@@ -58,12 +58,12 @@ namespace core
         return ::memcpy(dst, src, n);
     }
 
-    MLW_FORCE_INLINE void *mlwMemset(void *dst, u8 value, usize n)
+    MLW_FORCE_INLINE void *mlwMemset(void *dst, uint8 value, usize n)
     {
         return ::memset(dst, value, n);
     }
 
-    MLW_FORCE_INLINE void *mlwAlignedAlloc(usize size, u16 alignment)
+    MLW_FORCE_INLINE void *mlwAlignedAlloc(usize size, usize alignment)
     {
 #if defined(MLW_WINDOWS)
         return ::_aligned_malloc(size, alignment);
